% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aldist}
\alias{aldist}
\title{Align sequences and compute edit distance}
\usage{
aldist(seq1, seq2)
}
\arguments{
\item{seq1}{a vector with first sequences in the pair to align}

\item{seq2}{a vector with second sequences in the pair to align}
}
\value{
a vector with edit distances
}
\description{
Using edlib to quickly compute the edit distances between pairs of sequences
}
\author{
Jean Monlong
}
\keyword{internal}
