% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateOl.bipartite.R
\name{annotateOl.bipartite}
\alias{annotateOl.bipartite}
\title{Annotate the overlap between SVs using reciprocal overlap and bipartite clustering}
\usage{
annotateOl.bipartite(ol.gr, min.ol = 0.5)
}
\arguments{
\item{ol.gr}{overlaps prepared with \code{prepareOl} function.}

\item{min.ol}{the minimum overlap to be considered a potential match. Default is 0.5}
}
\value{
an updated and filtered version of the input ol.gr
}
\description{
Annotate the overlap between SVs using reciprocal overlap and bipartite clustering
}
\author{
Jean Monlong
}
\keyword{internal}
