% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateOl.coverage.R
\name{annotateOl.coverage}
\alias{annotateOl.coverage}
\title{Annotate the overlap between SVs using the reciprocal coverage method}
\usage{
annotateOl.coverage(ol.gr, min.cov = 0.5)
}
\arguments{
\item{ol.gr}{overlaps prepared with \code{prepareOl} function.}

\item{min.cov}{the minimum coverage to be considered a match. Default is 0.5}
}
\value{
an updated and filtered version of the input ol.gr
}
\description{
Annotate the overlap between SVs using the reciprocal coverage method
}
\author{
Jean Monlong
}
\keyword{internal}
