% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateOl.reciprocal.R
\name{annotateOl.reciprocal}
\alias{annotateOl.reciprocal}
\title{Annotate the overlap between SVs using the reciprocal overlap method}
\usage{
annotateOl.reciprocal(ol.gr, min.ol = 0.5)
}
\arguments{
\item{ol.gr}{overlaps prepared with \code{prepareOl} function.}

\item{min.ol}{the minimum overlap to be considered a potential match. Default is 0.5}
}
\value{
an updated and filtered version of the input ol.gr
}
\description{
Annotate the overlap between SVs using the reciprocal overlap method
}
\author{
Jean Monlong
}
\keyword{internal}
