% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInvSeq.R
\name{checkInvSeq}
\alias{checkInvSeq}
\title{Check if variants could be inversions}
\usage{
checkInvSeq(refseq, altseq)
}
\arguments{
\item{refseq}{the sequence of the reference allele}

\item{altseq}{the sequence of the alternate allele}
}
\value{
a boolean vector representing if the variants are inversions
}
\description{
Checks if the two sequences are more or less reverse complement.
The reverse complement of the ALT is aligned to the REF and return TRUE if it aligns >80%.
}
\author{
Jean Monlong
}
\keyword{internal}
