% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterSVs.R
\name{clusterSVs}
\alias{clusterSVs}
\title{Cluster SVs based on overlap/similarity}
\usage{
clusterSVs(
  svs.gr,
  min.rol = 0.8,
  max.ins.dist = 20,
  range.seq.comp = FALSE,
  ins.seq.comp = FALSE,
  simprep = NULL,
  nb.cores = 1,
  batch.maxsize = 5000,
  log.level = c("CRITICAL", "WARNING", "INFO")
)
}
\arguments{
\item{svs.gr}{A GRanges with SVs (for example read by \code{readSVvcf} or \code{readSVvcf.multisamps})}

\item{min.rol}{minimum reciprocal overlap for deletions and other "ranges" SVs. Default is 0.1}

\item{max.ins.dist}{maximum distance for insertions to be clustered.}

\item{range.seq.comp}{compare sequence instead of overlapping deletions/inversion/etc. Default is FALSE.}

\item{ins.seq.comp}{compare sequence instead of insertion sizes. Default is FALSE.}

\item{simprep}{optional simple repeat annotation. Default is NULL. If non-NULL, GRanges to be used to}

\item{nb.cores}{number of processors to use. Default is 1.}

\item{batch.maxsize}{batch size to aim. To reduce memory usage, see Details. Default is 5000.}

\item{log.level}{the level of information in the log. Default is "CRITICAL" (basically no log).}
}
\value{
the svs.gr object annotated with two new columns:
\item{svsite}{the ID of the cluster (or SV site)}
\item{clique}{is this cluster a clique, i.e. all SVs overlapping/matching all other SVs in the cluster}
}
\description{
Cluster SVs based on overlap/similarity
}
\details{
SVs are overlapped with each other. A graph is then built where nodes (SVs) are connected is
they overlap/match. A cluster is a component in this graph.

To reduce the memory usage, the SVs are first grossly clustered into batches. The actual clustering
(and graph construction) is then performed separately for each batch, potentially in parallel.
}
\examples{
\dontrun{

svs = readSVvcf('svs.vcf.gz', keep.ids=TRUE)
svs = clusterSVs(svs)

}
}
\author{
Jean Monlong
}
