% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countAlleles.R
\name{countAlleles}
\alias{countAlleles}
\title{Count alleles in SV sites across samples}
\usage{
countAlleles(vcf.file, sv.sites)
}
\arguments{
\item{vcf.file}{the path to the VCF file}

\item{sv.sites}{a list defining the SV sites (each element contains a vector with SV ids)}
}
\value{
a matrix with allele counts for each site (rows) and samples (columns)
}
\description{
Read a VCF file and count alleles for SVs grouped by SV site. The \code{sv.sites} parameter
is a list defining which SVs are assigned to which SV site. The SV ids are those
created by \code{readSVvcf.multisamps} which is expected to be run first.
}
\author{
Jean Monlong
}
