% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estTrimSize.R
\name{estTrimSize}
\alias{estTrimSize}
\title{Estimate the amount of sequence to right-trim}
\usage{
estTrimSize(seqs1, seqs2)
}
\arguments{
\item{seqs1}{a set of sequences, e.g. the reference sequences.}

\item{seqs2}{a set of sequences, e.g. the alt sequences.}
}
\value{
the number of bases to right trim.
}
\description{
Estimate the amount of sequence to right-trim
}
\author{
Jean Monlong
}
\keyword{internal}
