% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalOl.R
\name{evalOl}
\alias{evalOl}
\title{Compute FP, FN, TP}
\usage{
evalOl(ol.gr, truth.gr, calls.gr)
}
\arguments{
\item{ol.gr}{an GRanges with overlap information and filtered/annotated by \code{annotateOl}}

\item{truth.gr}{the original SVs from the truthset. Correspond to the queryHits in ol.gr. Use
only variants with 'pass' column.}

\item{calls.gr}{the original SVs from the callset. Correspond to the subjectHits in ol.gr. Use
only variants with 'pass' column.}
}
\value{
a list with
\item{eval}{a data.frame with FP, FN and TP for each SV type.}
\item{regions}{a list of FN/FP/TP regions for each SV type}
}
\description{
Compute FP, FN, TP
}
\author{
Jean Monlong
}
\keyword{internal}
