% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendSVwithSimpRep.R
\name{extendSVwithSimpRep}
\alias{extendSVwithSimpRep}
\title{Extend SV coordinates using simple repeat annotation}
\usage{
extendSVwithSimpRep(svs.gr, simprep.gr, max.dist.sr.join = 5, max.sv.dist = 5)
}
\arguments{
\item{svs.gr}{a GRanges object with SV information (e.g. read by \code{readSVvcf})}

\item{simprep.gr}{a GRanges object with simple repeat information.}

\item{max.dist.sr.join}{maximum distance to join nearby simple repeats. Default is 5 (bp)}

\item{max.sv.dist}{maximum distance between SV and simple repeat to apply the extension. Default is 5 (bp)}
}
\value{
an updated GRanges object for svs.gr (start/end extended by simprep.gr)
}
\description{
Equivalent SVs are sometimes recorded as quite different variants because placed at
different locations of a short tandem repeat. For example, imagine a large 100 bp
tandem repeat in the reference genome. An expansion of 50 bp might be represented
as a 50 bp insertion at the beginning of the repeat in the callset but at the end
of the repeat in the truth set. Because they are distant by 100 bp they might not
match. Instead of increasing the distance threshold too much, this function provides
a more flexible way of matching variants by first extending them with nearby simple
repeat. In this example, because we know of this tandem repeat, both insertions will
be extended to span the full annotated reference repeat, hence ensuring that they are
matched and compared (e.g. by reciprocal size or sequence alignment distance)
short tandem repeat.
}
\author{
Jean Monlong
}
\keyword{internal}
