% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterSVs.R
\name{filterSVs}
\alias{filterSVs}
\title{Filter SVs for regions of interest}
\usage{
filterSVs(sv.gr, regions.gr, ol.prop = 0.5)
}
\arguments{
\item{sv.gr}{the input SVs (e.g. read from \code{readSVvcf})}

\item{regions.gr}{the regions of interest}

\item{ol.prop}{minimum proportion of sv.gr that must overlap regions.gr.
Default is 0.5}
}
\value{
a subset of sv.gr that overlaps regions.gr
}
\description{
Filter SVs for regions of interest
}
\author{
Jean Monlong
}
