% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNocalls.R
\name{findNocalls}
\alias{findNocalls}
\title{Find no-calls variants}
\usage{
findNocalls(calls.gr, truth.gr, max.ins.dist = 20, min.cov = 0.5,
  min.del.rol = 0.1, ins.seq.comp = FALSE, nb.cores = 1,
  sample.name = NULL, check.inv = FALSE)
}
\arguments{
\item{calls.gr}{call set. A GRanges or the path to a VCF file.}

\item{truth.gr}{truth set. A GRanges or the path to a VCF file.}

\item{max.ins.dist}{maximum distance for insertions to be clustered. Default is 20.}

\item{min.cov}{the minimum coverage to be considered a match. Default is 0.5}

\item{min.del.rol}{minimum reciprocal overlap for deletions. Default is 0.1}

\item{ins.seq.comp}{compare sequence instead of insertion sizes. Default is FALSE.}

\item{nb.cores}{number of processors to use. Default is 1.}

\item{sample.name}{the name of the sample to use if VCF files given as
input. If NULL (default), use first sample.}

\item{check.inv}{should the sequence of MNV be compared to identify inversions.}
}
\value{
a data.frame with coordinates and variant ids from the truth set
corresponding to no-calls.
}
\description{
Compare calls with a truth set and identifies which variants from
the truth set specifically not called (genotype ./.).
}
\details{
Same overlapping strategy as in \code{svevalOl} although here no-calls
are kept and there is no splitting by genotype.
}
\author{
Jean Monlong
}
