% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivg_sv.R
\name{ivg_sv}
\alias{ivg_sv}
\title{Interactive exploration of SVs in a variation graph}
\usage{
ivg_sv(svs, xg, ucsc.genome = "hg38")
}
\arguments{
\item{svs}{either a GRanges with SVs (e.g. from \code{readSVvcf}) or the path to
a VCF file.}

\item{xg}{the path to the xg object of the variation graph.}

\item{ucsc.genome}{the genome version for the UCSC Genome Browser automated link.}
}
\value{
Starts a Shiny app in a web browser.
}
\description{
Opens a Shiny app with a dynamic table that contains input SVs. Clicking on a
SV (row in the table) generates a simplified representation of the variation
graph around this SV. The number of flanking nodes (context) can be increased
if necessary, e.g. for large insertions. vg needs to be installed
(https://github.com/vgteam/vg).
}
\author{
Jean Monlong
}
