% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{merge_ac_svsites_cpp}
\alias{merge_ac_svsites_cpp}
\title{Count alleles in SV sites across samples}
\usage{
merge_ac_svsites_cpp(filename, use_gz, sv_sites)
}
\arguments{
\item{filename}{the path to the VCF file (unzipped or gzipped).}

\item{use_gz}{is the VCF file gzipped?}

\item{sv_sites}{a list defining the variant ID for each sv site (element).}
}
\value{
matrix with allele counts for each sv site (rows) and sample (columns)
}
\description{
Count allele at the SV site level. This function will use variant ids
created by read_vcf_multisamps_cpp
}
\author{
Jean Monlong
}
\keyword{internal}
