% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olInsertions.R
\name{olInsertions}
\alias{olInsertions}
\title{Overlap calls and truth set for insertions}
\usage{
olInsertions(calls.gr, truth.gr, max.ins.gap = 1, ins.seq.comp = FALSE,
  nb.cores = 1, quiet = FALSE)
}
\arguments{
\item{calls.gr}{call set}

\item{truth.gr}{truth set}

\item{max.ins.gap}{maximum distance for insertions to be clustered.}

\item{ins.seq.comp}{compare sequence instead of insertion sizes. Default is FALSE.}

\item{nb.cores}{number of processors to use. Default is 1.}

\item{quiet}{if TRUE quiet mode with no messages.}
}
\value{
a list with
\item{calls}{the input calls}
\item{truth}{the input truth}
\item{ol}{a data.frame with the overlap information}
}
\description{
Overlap calls and truth set for insertions
}
\author{
Jean Monlong
}
\keyword{internal}
