% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olRanges.R
\name{olRanges}
\alias{olRanges}
\title{Overlap SVs as ranges (for deletions and inversions)}
\usage{
olRanges(calls.gr, truth.gr, min.rol = 0.1, type = c("DEL", "INV"))
}
\arguments{
\item{calls.gr}{the call set}

\item{truth.gr}{truth set}

\item{min.rol}{minimum reciprocal overlap for deletions (and inversions). Default is 0.1}

\item{type}{the type of SVs to overlap. 'DEL' (default) or 'INV'.}
}
\value{
a list with:
\item{calls}{the input calls}
\item{truth}{the input truth}
\item{rol.gr}{a GRanges with intersection of overlapping regions.}
}
\description{
Overlap SVs as ranges (for deletions and inversions)
}
\author{
Jean Monlong
}
\keyword{internal}
