% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_perregion.R
\name{plot_perregion}
\alias{plot_perregion}
\title{Recall, precision, F1 per region}
\usage{
plot_perregion(eval, regions.gr, min.region.ol = 0.5, plot = TRUE)
}
\arguments{
\item{eval}{the output of \code{svevalOl}.}

\item{regions.gr}{GRanges object with regions of interest}

\item{min.region.ol}{minimum proportion of variant that must overlap
regions.gr. Default is 0.5}

\item{plot}{should the function return the plot list. Default is TRUE. If FALSE,
returns a data.frame.}
}
\value{
a list of ggplot objects if plot=TRUE (default); a data.frame otherwise.
}
\description{
Recall, precision, F1 per region
}
\author{
Jean Monlong
}
