% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_persize.R
\name{plot_persize}
\alias{plot_persize}
\title{Recall, precision, F1 per SV size}
\usage{
plot_persize(
  eval,
  size.breaks = c(50, 100, 500, 1000, 10000, Inf),
  plot = TRUE
)
}
\arguments{
\item{eval}{the output of \code{svevalOl}.}

\item{size.breaks}{a vector for breaking the sizes into classes.}

\item{plot}{should the function return the plot list. Default is TRUE. If FALSE,
returns a data.frame.}
}
\value{
a list of ggplot objects if plot=TRUE (default); a data.frame otherwise.
}
\description{
Recall, precision, F1 per SV size
}
\author{
Jean Monlong
}
