% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prcurve.R
\name{plot_prcurve}
\alias{plot_prcurve}
\title{Create precision-recall graphs}
\usage{
plot_prcurve(eval, labels = NULL)
}
\arguments{
\item{eval}{a data.frame, a list of data.frames, or a vector with one or several paths to files with curve information.}

\item{labels}{the labels when multiple results are to be merged. Same length as the
list of vector. Ignored is NULL (default).}
}
\value{
list of ggplot graph objects
}
\description{
Create precision-recall graphs
}
\author{
Jean Monlong
}
