% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ranges.R
\name{plot_ranges}
\alias{plot_ranges}
\title{Plot variants in a region}
\usage{
plot_ranges(gr.l, region.gr = NULL, pt.size = 2, lab.size = 4,
  maxgap = 20, scale.legend = 50, show.svids = TRUE)
}
\arguments{
\item{gr.l}{a list of GRanges. If named, the names are used to name the graph's panel.}

\item{region.gr}{the region of interest. If NULL (default), all variants are displayed.}

\item{pt.size}{the point (and line) sizes. Default is 2.}

\item{lab.size}{the label size. Default is 4}

\item{maxgap}{the maximum gap allowed when filtering variants in regions. Default is 20.}

\item{scale.legend}{the size of the scale legend at the bottom. 0 to switch off. Default is 50 bp}

\item{show.svids}{should the SV ids be shown on the y-axis. Default is TRUE}
}
\value{
a ggplot2 object
}
\description{
A simple ggplot2 representation of variants in a region.
The beginning of the variant is represented as a point (shape=SV type). The point is
annotated with the variant size. A line outlines the range (e.g. for deletions or
inversions).
}
\author{
Jean Monlong
}
