% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prf.R
\name{prf}
\alias{prf}
\title{Compute precision, recall and F1 score}
\usage{
prf(eval.df)
}
\arguments{
\item{eval.df}{a data.frame with columns TP, TP.baseline, FP, and FN.}
}
\value{
the input data.frame with 3 new columns precision, recall and F1.
}
\description{
Compute the precision, recall and F1 score using the TP, TP.baseline, FP
and FN columns.
}
\author{
Jean Monlong
}
