% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_vcf_cpp}
\alias{read_vcf_cpp}
\title{Read VCF using CPP reader}
\usage{
read_vcf_cpp(
  filename,
  use_gz,
  sample_name = "",
  min_sv_size = 10L,
  shorten_ref = TRUE,
  shorten_alt = TRUE,
  gq_field = "GQ",
  check_inv = FALSE,
  keep_nocalls = FALSE,
  other_field = ""
)
}
\arguments{
\item{filename}{the path to the VCF file (unzipped or gzipped).}

\item{use_gz}{is the VCF file gzipped?}

\item{sample_name}{which sample to process. If not found, uses first sample in VCF file.
If "*", force no sample selection}

\item{min_sv_size}{minimum variant size to keep in bp. Variants shorter than this
will be skipped. Default is 10.}

\item{shorten_ref}{should the REF sequence be shortened to the first 10 bp. Default is TRUE}

\item{shorten_alt}{should the ALT sequence be shortened to the first 10 bp. Default is TRUE}

\item{gq_field}{which field from FORMAT should be used as genotype quality. Default is "GQ".
If not found, QUAL will be used}

\item{check_inv}{guess if a variant is an inversion by aligning REF with the
reverse complement of ALT. If >80\% similar (and REF and ALT>10bp), variant is classified as INV.}

\item{keep_nocalls}{should we keep variants/alleles with missing genotypes (e.g. "./.").
Default is FALSE}

\item{other_field}{name of another field from INFO to extract.}
}
\value{
data.frame with variant and genotype information
}
\description{
For each VCF record the information in the INFO field is used in priority. If
missing information is guessed from the REF/ALT sequences.
If multiple alleles are defined in ALT, they are split and the allele count extracted
from the GT field.
}
\details{
Alleles are split and, for each, column 'ac' reports the allele count. Notable cases incude
'ac=-1' for no/missing calls (e.g. './.'), and 'ac=0' on the first allele to report hom ref,
variants. These cases are often filtered later with 'ac>0' to keep only non-ref calls. If
the VCF contains no samples or if no sample selection if forced (sample_name='*'), 'ac' will
contain '-1' for all variants in the VCF.
}
\author{
Jean Monlong
}
\keyword{internal}
