% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSVvcf.multisamps.R
\name{readSVvcf.multisamps}
\alias{readSVvcf.multisamps}
\title{Read SVs from a VCF file}
\usage{
readSVvcf.multisamps(vcf.file, keep.ins.seq = FALSE,
  keep.ref.seq = FALSE, check.inv = FALSE, keep.ids = FALSE,
  out.fmt = c("gr", "df"), min.sv.size = 10)
}
\arguments{
\item{vcf.file}{the path to the VCF file}

\item{keep.ins.seq}{should it keep the inserted sequence? Default is FALSE.}

\item{keep.ref.seq}{should it keep the reference allele sequence? Default is FALSE.}

\item{check.inv}{should the sequence of MNV be compared to identify inversions.}

\item{keep.ids}{keep variant ids? Default is FALSE.}

\item{out.fmt}{output format. Default is 'gr' for GRanges. Other options: 'df' for
data.frame.}

\item{min.sv.size}{the minimum size of the variant to extract from the VCF. Default is 10}
}
\value{
depending on 'out.fmt', a GRanges or a data.frame with relevant information.
}
\description{
Read a VCF file that contains SVs for multiple samples and create a GRanges with population estimates.
}
\details{
Alleles are split and, for each, column 'ac' reports the total allele count. The number of "ref" genotypes,
i.e. homozygous refs e.g. '0/0', and the number of called samples (not missing './.') are also counted.
The 'af' column contains an estimate of the allele frequency.
}
\examples{
\dontrun{
svs.gr = readSVvcf.multisamps('svs.vcf')
}
}
\author{
Jean Monlong
}
