% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmskAnnotate.R
\name{rmskAnnotate}
\alias{rmskAnnotate}
\title{Annotate REF/ALT sequence with RepeatMasker}
\usage{
rmskAnnotate(svs.gr, nb.cores = 1, species = "human")
}
\arguments{
\item{svs.gr}{SVs. A GRanges or the path to a VCF file.}

\item{nb.cores}{the number of cores that RepeatMasker should use. Default:1.}

\item{species}{the species to use in RepeatMasker. Default: human.}
}
\value{
an updated GRanges with new columns 'rmsk.name', 'rmsk.classfam' and 'rmsk.cov'.
}
\description{
Extracts REF/ALT sequence from a VCF, runs RepeatMasker to annotate transposable
elements and simple repeats, and annotates the original variants.
}
\details{
This is a simple annotation where only the main repeat class is retrieved for
each variant (covering most of the sequence).

RepeatMasker must be installed.
}
\examples{
\dontrun{
svs.gr = readSVvcf('calls.vcf', keep.ins.seq=TRUE, keep.ref.seq=TRUE)
svs.gr = rmskAnnotate(svs.gr)
}
}
\author{
Jean Monlong
}
