% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitchMergeHets.R
\name{stitchMergeHets}
\alias{stitchMergeHets}
\title{Stitch and merge heterozygous SVs}
\usage{
stitchMergeHets(
  svs.gr,
  do.stitch = TRUE,
  do.merge = TRUE,
  stitch.dist = 20,
  min.rol = 0.8,
  max.ins.dist = 20,
  range.seq.comp = FALSE,
  ins.seq.comp = FALSE
)
}
\arguments{
\item{svs.gr}{input GRanges with SVs, e.g. read from \code{readSVvcf}.}

\item{do.stitch}{should nearby het SVs of the same type be stitched? Default is TRUE.}

\item{do.merge}{should similar het SVs of the same type "merged" into one homozygous variant? Default is TRUE.}

\item{stitch.dist}{the maximum distance allowed for two SVs to be stitched.}

\item{min.rol}{minimum reciprocal overlap to merge two hets into one hom}

\item{max.ins.dist}{maximum distance for insertions to be clustered when merging hets.}

\item{range.seq.comp}{compare sequence instead of overlapping deletions/inversion/etc. Default is FALSE.}

\item{ins.seq.comp}{compare sequence instead of insertion sizes. Default is FALSE.}
}
\value{
a GRanges with het SVs from input svs.gr stitched and/or merged.
}
\description{
Fragmented calls often can't be matched as is to another call set.
Calls that are fragmented in multiple pieces could be stitched together if they
are close by and with the same genotype and SV type.
An homozygous call can also be fragmented in two heterozygous calls. To counter this,
we can merge pairs of heterozygous SVs (from the same type) that overlap substantially.
}
\author{
Jean Monlong
}
