% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitchSVs.R
\name{stitchSVs}
\alias{stitchSVs}
\title{Stitch fragmented SVs}
\usage{
stitchSVs(svs, stitch.dist = 20)
}
\arguments{
\item{svs}{GRanges with SV information}

\item{stitch.dist}{the maximum distance allowed for two SVs to be stitched.}
}
\value{
an upated GRanges.
}
\description{
Stitch together SVs that are close together. This should be done for SVs of the same
type and genotype. When several SV pairs can be stitched in a cluster, only the best
pair (closest to each other) is merged, hence this could be run several times (e.g. until the
size of the output is similar to the input's).
}
\author{
Jean Monlong
}
\keyword{internal}
