% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_eval.R
\name{subset_eval}
\alias{subset_eval}
\title{Compute evaluation results on a subset of the calls}
\usage{
subset_eval(
  eval,
  regions.gr = NULL,
  accepted.filters = NULL,
  min.region.ol = 0.5,
  nb.cores = 1
)
}
\arguments{
\item{eval}{the output of \code{svevalOl}.}

\item{regions.gr}{GRanges object with regions of interest. NULL (default) means all SVs.}

\item{accepted.filters}{vector of the values of the FILTER field to keep. If NULL (default), all values are accepted
which means all values are kept}

\item{min.region.ol}{minimum proportion of variant that must overlap
regions.gr. Default is 0.5}

\item{nb.cores}{number of processors to use. Default is 1.}
}
\value{
a list like for \code{svevalOl}
}
\description{
Compute evaluation results on a subset of the calls
}
\author{
Jean Monlong
}
