% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sveval-package.R
\docType{package}
\name{sveval-package}
\alias{sveval-package}
\title{SV evaluation}
\description{
Evaluate SV in a call set against a truth set using overlap-based approaches and sequence comparison for insertions.
}
\details{
\tabular{ll}{
Package: \tab sveval\cr
Type: \tab Package\cr
Version: \tab 2.2.0\cr
Date: \tab 2019-09-16\cr
License: \tab MIT\cr
}
}
\examples{
\dontrun{
eval = svevalOl('calls.vcf', 'truth.vcf')
plot_prcurve(eval$curve)

# Comparing multiple methods
eval.1 = svevalOl('calls1.vcf', 'truth.vcf')
eval.2 = svevalOl('calls2.vcf', 'truth.vcf')
plot_prcurve(list(eval.1$curve, eval.2$curve), labels=c('method1', 'method2'))
}
}
\seealso{
\url{http://www.github.com/jmonlong/sveval}
}
\author{
Jean Monlong \email{jmonlong@ucsc.edu}
}
