% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svevalOl.R
\name{svevalOl}
\alias{svevalOl}
\title{SV evaluation based on overlap and variant size}
\usage{
svevalOl(calls.gr, truth.gr, max.ins.dist = 20, min.cov = 0.5,
  min.del.rol = 0.1, ins.seq.comp = FALSE, nb.cores = 1,
  min.size = 50, max.size = Inf, bed.regions = NULL,
  bed.regions.ol = 0.5, qual.field = c("QUAL", "GQ"),
  sample.name = NULL, outfile = NULL, out.bed.prefix = NULL,
  qual.quantiles = seq(0, 1, 0.1), check.inv = FALSE,
  geno.eval = FALSE, stitch.hets = FALSE, stitch.dist = 20,
  merge.hets = FALSE, merge.rol = 0.8)
}
\arguments{
\item{calls.gr}{call set. A GRanges or the path to a VCF file.}

\item{truth.gr}{truth set. A GRanges or the path to a VCF file.}

\item{max.ins.dist}{maximum distance for insertions to be clustered. Default is 20.}

\item{min.cov}{the minimum coverage to be considered a match. Default is 0.5}

\item{min.del.rol}{minimum reciprocal overlap for deletions. Default is 0.1}

\item{ins.seq.comp}{compare sequence instead of insertion sizes. Default is FALSE.}

\item{nb.cores}{number of processors to use. Default is 1.}

\item{min.size}{the minimum SV size to be considered. Default 50.}

\item{max.size}{the maximum SV size to be considered. Default is Inf.}

\item{bed.regions}{If non-NULL, a GRanges object or path to a BED file
(no headers) with regions of interest.}

\item{bed.regions.ol}{minimum proportion of sv.gr that must overlap
regions.gr. Default is 0.5}

\item{qual.field}{fields to use as quality. Will be tried in order.}

\item{sample.name}{the name of the sample to use if VCF files given as
input. If NULL (default), use first sample.}

\item{outfile}{the TSV file to output the results. If NULL (default), returns a data.frame.}

\item{out.bed.prefix}{prefix for the output BED files. If NULL (default), no BED output.}

\item{qual.quantiles}{the QUAL quantiles for the PR curve. Default is (0, .1, ..., .9, 1).}

\item{check.inv}{should the sequence of MNV be compared to identify inversions.}

\item{geno.eval}{should het/hom be evaluated separately (genotype evaluation). Default
FALSE.}

\item{stitch.hets}{should clustered hets be stitched together before genotype evatuation.
Default is FALSE.}

\item{stitch.dist}{the maximum distance to stitch hets during genotype evaluation.}

\item{merge.hets}{should similar hets be merged into homs before genotype evaluation.
Default is FALSE.}

\item{merge.rol}{the minimum reciprocal overlap to merge hets before genotype
evaluation.}
}
\value{
a list with
\item{eval}{a data.frame with TP, FP and FN for each SV type when including all variants}
\item{curve}{a data.frame with TP, FP and FN for each SV type when using different quality thesholds}
}
\description{
SV evaluation based on overlap and variant size
}
\examples{
\dontrun{
## From VCF files
eval = svevalOl('calls.vcf', 'truth.vcf')

## From GRanges
calls.gr = readSVvcf('calls.vcf')
truth.gr = readSVvcf('truth.vcf')
eval = svevalOl(calls.gr, truth.gr)

## Genotype evaluation
eval = svevalOl(calls.gr, truth.gr, geno.eval=TRUE, merge.hets=TRUE, stitch.hets=TRUE)
}
}
\author{
Jean Monlong
}
