% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svOverlap.R
\name{svOverlap}
\alias{svOverlap}
\title{Overlap and annotate SV sets with coverage metrics}
\usage{
svOverlap(
  query,
  subject,
  max.ins.dist = 20,
  min.cov = 0.5,
  min.del.rol = 0.1,
  ins.seq.comp = FALSE,
  nb.cores = 1
)
}
\arguments{
\item{query}{a GRanges object with SVs}

\item{subject}{another GRanges object with SVs}

\item{max.ins.dist}{maximum distance for insertions to be clustered. Default is 20.}

\item{min.cov}{the minimum coverage to be considered a match. Default is 0.5}

\item{min.del.rol}{minimum reciprocal overlap for deletions. Default is 0.1}

\item{ins.seq.comp}{compare sequence instead of insertion sizes. Default is FALSE.}

\item{nb.cores}{number of processors to use. Default is 1.}
}
\value{
a list with:
\item{query}{the query GRanges object annotated}
\item{subject}{the subject GRanges object annotated}
}
\description{
Overlap and annotate SV sets with coverage metrics
}
\author{
Jean Monlong
}
