% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_mergetsvs.R
\name{wrapper_mergetsvs}
\alias{wrapper_mergetsvs}
\title{Internal wrapper function to merge TSVs as a command line tool}
\usage{
wrapper_mergetsvs(args)
}
\arguments{
\item{args}{arguments}
}
\value{
return code (0: success, 1: error)
}
\description{
Wrapper functions that merges TSV files made by sveval and make plots
}
\author{
Jean Monlong
}
\keyword{internal}
