% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_sveval.R
\name{wrapper_sveval}
\alias{wrapper_sveval}
\title{Internal wrapper function to run sveval as a command line tool}
\usage{
wrapper_sveval(args)
}
\arguments{
\item{args}{arguments}
}
\value{
return code (0: success, 1: error)
}
\description{
Wrapper functions that reads arguments and runs sveval.
}
\author{
Jean Monlong
}
\keyword{internal}
