% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateOl.R
\name{annotateOl}
\alias{annotateOl}
\title{Annotate SVs from overlap}
\usage{
annotateOl(ol.l, min.qual = 0, method = c("coverage", "bipartite"))
}
\arguments{
\item{ol.l}{output of an overlap function (olInsertions or olRanges).}

\item{min.qual}{the minimum quality considered for the calls.}

\item{method}{the method to annotate the overlap. Either 'coverage' (default) for the
cumulative coverage (e.g. to deal with fragmented calls); or 'bipartite' for a 1-to-1
matching of variants in the calls and truth sets.}
}
\value{
an updated list with a *cov* column added to the calls and truth sets.
}
\description{
Annotate SVs from overlap
}
\author{
Jean Monlong
}
\keyword{internal}
