% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prf.R
\name{prf}
\alias{prf}
\title{Compute precision, recall and F1 score}
\usage{
prf(eval.df, use.calls.tp.everywhere = TRUE)
}
\arguments{
\item{eval.df}{a data.frame with columns TP, TP.baseline, FP, and FN.}

\item{use.calls.tp.everywhere}{use the TP count as number of calls matched even for the recall computation}
}
\value{
the input data.frame with 3 new columns precision, recall and F1.
}
\description{
Compute the precision, recall and F1 score using the TP, TP.baseline, FP
and FN columns.
}
\author{
Jean Monlong
}
