% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSVvcf.R
\name{readSVvcf}
\alias{readSVvcf}
\title{Read SVs from a VCF file}
\usage{
readSVvcf(vcf.file, keep.ins.seq = FALSE, sample.name = NULL)
}
\arguments{
\item{vcf.file}{the path to the VCF file}

\item{keep.ins.seq}{should it keep the inserted sequence? Default is FALSE.}

\item{sample.name}{the name of the sample to use. If NULL (default), use
first sample.}
}
\value{
a GRanges object with relevant information.
}
\description{
Read SVs from a VCF file
}
\examples{
\dontrun{
calls.gr = readSVvcf('calls.vcf')
}
}
\author{
Jean Monlong
}
