% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSVvcf.R
\name{readSVvcf}
\alias{readSVvcf}
\title{Read SVs from a VCF file}
\usage{
readSVvcf(vcf.file, keep.ins.seq = FALSE, keep.ref.seq = FALSE,
  sample.name = NULL, qual.field = c("GQ", "QUAL"),
  check.inv = FALSE, keep.ids = FALSE, nocalls = FALSE,
  right.trim = TRUE, vcf.object = FALSE)
}
\arguments{
\item{vcf.file}{the path to the VCF file}

\item{keep.ins.seq}{should it keep the inserted sequence? Default is FALSE.}

\item{keep.ref.seq}{should it keep the reference allele sequence? Default is FALSE.}

\item{sample.name}{the name of the sample to use. If NULL (default), use
first sample.}

\item{qual.field}{fields to use as quality. Will be tried in order.}

\item{check.inv}{should the sequence of MNV be compared to identify inversions.}

\item{keep.ids}{keep variant ids? Default is FALSE.}

\item{nocalls}{if TRUE returns no-calls only (genotype ./.). Default FALSE.}

\item{right.trim}{if TRUE (default) the REF/ALT sequences are right-trimmed
after splitting up multi-ALT variants.}

\item{vcf.object}{should the output be a VCF object instead. Default is FALSE.}
}
\value{
a GRanges object with relevant information.
}
\description{
Read a VCF file that contains SVs and create a GRanges with relevant information, e.g. SV size or genotype quality.
}
\details{
By default, the quality information is taken from the QUAL field. If all
values are NA or 0, the function will try other fields as speficied in
the "qual.field" vector. Fields can be from the INFO or FORMAT fields.
}
\examples{
\dontrun{
calls.gr = readSVvcf('calls.vcf')
}
}
\author{
Jean Monlong
}
