% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSVvcf.R
\name{readSVvcf}
\alias{readSVvcf}
\title{Read SVs from a VCF file}
\usage{
readSVvcf(vcf.file, keep.ins.seq = FALSE, sample.name = NULL,
  qual.field = c("QUAL", "GQ"), check.inv = FALSE)
}
\arguments{
\item{vcf.file}{the path to the VCF file}

\item{keep.ins.seq}{should it keep the inserted sequence? Default is FALSE.}

\item{sample.name}{the name of the sample to use. If NULL (default), use
first sample.}

\item{qual.field}{fields to use as quality. Will be tried in order.}

\item{check.inv}{should the sequence of MNV be compared to identify inversions.}
}
\value{
a GRanges object with relevant information.
}
\description{
Read a VCF file that contains SVs and create a GRanges with relevant information, e.g. SV size or genotype quality.
}
\details{
By default, the quality information is taken from the QUAL field. If all
values are NA or 0, the function will try other fields as speficied in
the "qual.field" vector. Fields can be from the INFO or FORMAT fields.
}
\examples{
\dontrun{
calls.gr = readSVvcf('calls.vcf')
}
}
\author{
Jean Monlong
}
